+++ resources/hpxml-measures/workflow/tests/hpxml_translator_test.rb
@@ -591,8 +591,8 @@ class HPXMLTest < MiniTest::Test
     if not xml.include? 'ASHRAE_Standard_140'
       sum_component_htg_loads = results.select { |k, v| k.start_with? 'Component Load: Heating:' }.map { |k, v| v }.sum(0.0)
       sum_component_clg_loads = results.select { |k, v| k.start_with? 'Component Load: Cooling:' }.map { |k, v| v }.sum(0.0)
-      residual_htg_load = results['Load: Heating (MBtu)'] - sum_component_htg_loads
-      residual_clg_load = results['Load: Cooling (MBtu)'] - sum_component_clg_loads
+      residual_htg_load = results['Load: Heating: Delivered (MBtu)'] - sum_component_htg_loads
+      residual_clg_load = results['Load: Cooling: Delivered (MBtu)'] - sum_component_clg_loads
       assert_operator(residual_htg_load.abs, :<, 0.5)
       assert_operator(residual_clg_load.abs, :<, 0.5)
     end
@@ -1497,7 +1497,7 @@ class HPXMLTest < MiniTest::Test
       all_results.sort.each do |xml, xml_results|
         next unless xml.include? 'C.xml'
 
-        htg_load = xml_results['Load: Heating (MBtu)'].round(2)
+        htg_load = xml_results['Load: Heating: Delivered (MBtu)'].round(2)
         csv << [File.basename(xml), htg_load, 'N/A']
         test_name = File.basename(xml, File.extname(xml))
         htg_loads[test_name] = htg_load
@@ -1505,7 +1505,7 @@ class HPXMLTest < MiniTest::Test
       all_results.sort.each do |xml, xml_results|
         next unless xml.include? 'L.xml'
 
-        clg_load = xml_results['Load: Cooling (MBtu)'].round(2)
+        clg_load = xml_results['Load: Cooling: Delivered (MBtu)'].round(2)
         csv << [File.basename(xml), 'N/A', clg_load]
         test_name = File.basename(xml, File.extname(xml))
         clg_loads[test_name] = clg_load